Fix current leak during deep sleep
Unusually high power consumption (10-20ma) was observed when the system
was put in DEEP_SLEEP mode after having previously been operational.
The cause was that the pins controlling the data on the LCD screen were
left in an undefined state after turning the screen on. As the Nokia_LCD
class neither handles the screen when completely powered off nor cares for
the state of the pins, its user has to set the state of those pins themselves.
This could possibly be an interesting addition to the Nokia_LCD library,
i.e. handling the power as well as the backlight pins.